Flip mesh texcoords in place and move texture fixture to conftest
